chore: rename gloglog to goldlog
