+++ packages/iceworks-client/src/pages/Project/components/GitPanel/index.js
@@ -4,7 +4,7 @@ import { injectIntl, FormattedMessage } from 'react-intl';
 import { Message } from '@alifd/next';
 import Modal from '@components/Modal';
 import useModal from '@hooks/useModal';
-import glodlog from '@utils/glodlog';
+import goldlog from '@utils/goldlog';
 import Panel from '../Panel';
 import PanelHead from '../Panel/head';
 import GitRemote from './GitRemote';
@@ -41,7 +41,7 @@ const GitPanel = ({ intl, title, description }) => {
   async function onInit(setRemoteUrl) {
     await gitStore.init(setRemoteUrl);
     await gitStore.refresh();
-    glodlog('git-init', {
+    goldlog('git-init', {
       type: 'project',
       url: setRemoteUrl,
     });
