+++ packages/iceworks-client/src/pages/Project/components/MenuPanel/index.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage, injectIntl } from 'react-intl';
 import { Tab } from '@alifd/next';
-import glodlog from '@utils/glodlog';
+import goldlog from '@utils/goldlog';
 import useModal from '@hooks/useModal';
 import cloneDeep from 'lodash.clonedeep';
 import MenuTreeConfig from '../../../../components/MenuTreeConfig';
@@ -78,7 +78,7 @@ const MenuPanel = ({ intl, title, description }) => {
       }, true);
     }
     await onChangeTree(copyData);
-    glodlog('create-menu', {
+    goldlog('create-menu', {
       type: 'project',
     });
   }
