+++ packages/iceworks-client/src/pages/Project/components/PagePanel/BuildPageModal.js
@@ -16,7 +16,7 @@ import {
   SortableHandle,
   arrayMove,
 } from 'react-sortable-hoc';
-import glodlog from '@utils/glodlog';
+import goldlog from '@utils/goldlog';
 import SavePageModal from './SavePageModal';
 import styles from './BuildPageModal.module.scss';
 
@@ -214,7 +214,7 @@ const BuildPageModal = ({
       ...data,
     });
     await progress.hide();
-    glodlog('create-page', {
+    goldlog('create-page', {
       type: 'project',
     });
     setSaveModal(false);
