+++ packages/iceworks-client/src/pages/Project/components/RouterPanel/index.js
@@ -4,7 +4,7 @@ import { injectIntl, FormattedMessage } from 'react-intl';
 import useModal from '@hooks/useModal';
 import { Message, Balloon } from '@alifd/next';
 import Icon from '@components/Icon';
-import glodlog from '@utils/glodlog'
+import goldlog from '@utils/goldlog'
 
 import CreateRouterModal from './CreateRouterModal';
 import DeleteRouterModal from './DeleteRouterModal';
@@ -98,7 +98,7 @@ const RouterPanel = ({ intl, title, description }) => {
       }
     }
     await onChangeData(dataSource);
-    glodlog('create-router', {
+    goldlog('create-router', {
       type: 'project',
     });
   }
